Restructure in folders/namespaces
