+++ src/Umbraco.StorageProviders.AzureBlob/AzureBlobMediaMiddleware.cs
@@ -13,6 +13,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Options;
 using Microsoft.Net.Http.Headers;
 using Umbraco.Cms.Core.Hosting;
+using Umbraco.StorageProviders.AzureBlob.IO;
 
 namespace Umbraco.StorageProviders.AzureBlob
 {
