+++ src/Umbraco.StorageProviders.AzureBlob/IO/IAzureBlobFileSystem.cs
@@ -1,7 +1,7 @@
 using Azure.Storage.Blobs;
 using Umbraco.Cms.Core.IO;
 
-namespace Umbraco.StorageProviders.AzureBlob
+namespace Umbraco.StorageProviders.AzureBlob.IO
 {
     /// <summary>
     /// The Azure Blob File System.
